// packages/web/src/generated/types/loanStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum LoanStatus {
  Pending,
  Active,
  Repaid,
  Cancelled,
}

export type LoanStatusArgs = LoanStatus;

export function getLoanStatusEncoder(): FixedSizeEncoder<LoanStatusArgs> {
  return getEnumEncoder(LoanStatus);
}

export function getLoanStatusDecoder(): FixedSizeDecoder<LoanStatus> {
  return getEnumDecoder(LoanStatus);
}

export function getLoanStatusCodec(): FixedSizeCodec<
  LoanStatusArgs,
  LoanStatus
> {
  return combineCodec(getLoanStatusEncoder(), getLoanStatusDecoder());
}